+++ JarvisAI/JarvisAI/JarvisAI/features/chatbot/.google-cookie
@@ -1,0 +1,3 @@
+#LWP-Cookies-2.0
+Set-Cookie3: 1P_JAR="2021-09-11-07"; path="/"; domain=".google.com"; path_spec; domain_dot; secure; expires="2021-10-11 07:48:40Z"; version=0
+Set-Cookie3: NID="223=cvjn9INyEVmeM-8b4qC3CNiO3JUNZLbPHNWt6ryE4PLCnbfurts8QDR6xnG7Soi3YDbehh3z_wpbwqrJFQFQqPXTj_MfpVgwWUcA2Hz68hLuJJ29gcaR5TfOGlpg0lgWj1z2Ru6mGUPt9ROzEFSQMntt5klm0jT2GNAvWPTx1pA"; path="/"; domain=".google.com"; path_spec; domain_dot; expires="2022-03-13 07:48:40Z"; HttpOnly=None; version=0
