+++ JarvisAI/JarvisAI/JarvisAI/features/face_recognition/predict.py
