+++ JarvisAI/JarvisAI/JarvisAI/features/hot_word_detection/hot_word_detection.py
