+++ JarvisAI/JarvisAI/JarvisAI/features/launch_app/launch_app.py
