+++ JarvisAI/JarvisAI/requirements.txt
@@ -33,7 +33,7 @@ lxml==4.6.1
 Markdown==3.3.3
 matplotlib==3.3.4
 oauthlib==3.1.0
-opencv-contrib-python==4.5.1.48
+opencv-contrib-python
 opt-einsum==3.3.0
 packaging==20.9
 Pillow==8.1.0
