+++ JarvisAI/cmd_twine.txt
@@ -1,0 +1,2 @@
+python setup.py sdist bdist_wheel
+twine upload dist/*
